input: Fix possible errors in Lite-On LTR-558 Sensor driver
PVS-Studio:
V547 Expression 'low < 0' is always false. Unsigned type value is never < 0
V547 Expression 'high < 0' is always false. Unsigned type value is never < 0

+++ drivers/input/misc/ltr558.c
@@ -647,7 +647,7 @@ static ssize_t ltr558_store_debug_regs(struct device *dev,
 static ssize_t ltr558_show_adc_data(struct device *dev,
 		struct device_attribute *attr, char *buf)
 {
-	u8 high,low;
+	int high,low;
 	char *after;
 	after = buf;
